requirements.txt
streamlit
pandas
folium
streamlit-folium
joblib
azure-storage-blob